% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_clean_styles.R
\name{cyto_clean_styles}
\alias{cyto_clean_styles}
\title{Cleaning nearly all styles in cytoscape session and import simple styles}
\usage{
cyto_clean_styles()
}
\value{
Cytoscape cleaned of styles and only styles with white and grey nodes.
}
\description{
Cleaning nearly all styles in cytoscape session and import simple styles
}
\examples{
\donttest{
cyto_clean_styles()
}

}
