% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_create_cpm_style.R
\name{cyto_create_cpm_style}
\alias{cyto_create_cpm_style}
\title{Create CPM style in Cytoscape}
\usage{
cyto_create_cpm_style(graph_input, k = 3, style_name = "auto")
}
\description{
Function to
}
