% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendroNetwork-package.R
\docType{package}
\name{dendroNetwork-package}
\alias{dendroNetwork}
\alias{dendroNetwork-package}
\title{dendroNetwork: Create networks of dendrochronological series using pairwise similarity}
\description{
This package makes it easy to create dendrochronological networks based on the similarity between tree-ring series or chronologies. It includes various functions to compare tree-ring curves building upon dplR. The networks can be used to visualise and understand the relations between tree-ring curves. These networks are also very useful to estimate the provenance or wood-use within a structure/context/site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RonaldVisser/dendroNetwork}
  \item \url{https://ronaldvisser.github.io/dendroNetwork/}
  \item Report bugs at \url{https://github.com/RonaldVisser/dendroNetwork/issues}
}

}
\author{
\strong{Maintainer}: Ronald Visser \email{r.m.visser@saxion.nl} (\href{https://orcid.org/0000-0001-6966-1729}{ORCID})

Other contributors:
\itemize{
  \item Angelino Salentino (\href{https://orcid.org/0000-0002-4763-3943}{ORCID}) [contributor]
  \item Andy Bunn (\href{https://orcid.org/0000-0001-9027-2162}{ORCID}) [contributor]
  \item Kaija Gahm (Kaija Gahm reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/627#issuecomment-1993235710) [reviewer]
  \item Zachary Gajewski (Zachary Gajewski reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/627#issuecomment-2037644081) [reviewer]
}

}
\keyword{internal}
