% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_table.R
\name{sim_table}
\alias{sim_table}
\title{Similarity table of tree-ring curves for the creation of a network}
\usage{
sim_table(trs1, trs2 = NULL, min_overlap = 50, last_digit_radius = FALSE)
}
\arguments{
\item{trs1}{Rwl object with first tree-ring series to be compared with trs2}

\item{trs2}{Optional second rwl object with second tree-ring series to be compared with trs1. Use this is you have to datasets that you want to compare.}

\item{min_overlap}{If the overlap of the compared series is longer or equal than this minimal value, the similarities will be calculated for the comparison}

\item{last_digit_radius}{Set this to TRUE if the last digit of a series name is the radius of the tree-ring series}
}
\value{
a list with all comparisons of the tree-ring series in trs1 (or between trs1 and trs2). This list includes the overlap, correlation (both with and without Hollstein-transformation), the t-value based on these correlations, the SGC, SSGC and the related probability of exceedence.
}
\description{
Function to calculate various similarity measures for the creation of dendrochronological networks as described by Visser (2022)....
}
\details{
[check this a list] with all comparisons of the tree-ring series in trs1 (or between trs1 and trs2). This list includes the overlap, correlation (both with and without Hollstein-transformation), the t-value based on these correlations, the SGC, SSGC and the related probability of exceedence.
}
\examples{
data(anos1)
sim_table(anos1)
sim_table(anos1, min_overlap = 25)


}
