% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_value.R
\name{t_value}
\alias{t_value}
\title{Students' t value}
\usage{
t_value(r, n)
}
\arguments{
\item{r}{correlation value.}

\item{n}{number over overlapping tree-rings/compared values.}
}
\value{
Students' t value as a numeric.
}
\description{
Function to determine the Students' t-value using the correlation and the number of compare values (tree-rings)
}
\examples{
t_value(0.5, 100)
}
