% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendroNetwork-package.R
\docType{package}
\name{dendroNetwork-package}
\alias{dendroNetwork}
\alias{dendroNetwork-package}
\title{dendroNetwork: Create Networks of Dendrochronological Series using Pairwise Similarity}
\description{
Creating dendrochronological networks based on the similarity between tree-ring series or chronologies. The package includes various functions to compare tree-ring curves building upon the 'dplR' package. The networks can be used to visualise and understand the relations between tree-ring curves. These networks are also very useful to estimate the provenance of wood as described in Visser (2021) \doi{10.5334/jcaa.79} or wood-use within a structure/context/site as described in Visser and Vorst (2022) \doi{10.1163/27723194-bja10014}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/dendroNetwork}
  \item \url{https://ropensci.github.io/dendroNetwork/}
  \item Report bugs at \url{https://github.com/ropensci/dendroNetwork/issues}
}

}
\author{
\strong{Maintainer}: Ronald Visser \email{r.m.visser@saxion.nl} (\href{https://orcid.org/0000-0001-6966-1729}{ORCID})

Other contributors:
\itemize{
  \item Angelino Salentino (\href{https://orcid.org/0000-0002-4763-3943}{ORCID}) [contributor]
  \item Andy Bunn (\href{https://orcid.org/0000-0001-9027-2162}{ORCID}) [contributor]
  \item Kaija Gahm (Kaija Gahm reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/627#issuecomment-1993235710) [reviewer]
  \item Zachary Gajewski (Zachary Gajewski reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/627#issuecomment-2037644081) [reviewer]
}

}
\keyword{internal}
